refactor : clean up file names and pakages
