+++ src/main/java/nextstep/domain/theme/Theme.java
@@ -1,4 +1,4 @@
-package nextstep;
+package nextstep.domain.theme;
 
 public class Theme {
     private String name;
@@ -23,9 +23,9 @@ public class Theme {
         return price;
     }
 
-    @Override
-    public boolean equals(Object obj) {
-        Theme theme = (Theme) obj;
-        return this.name.equals(theme.getName());
+    public boolean equals(Theme theme) {
+        return this.name.equals(theme.getName())
+                && this.desc.equals(theme.getDesc())
+                && this.price.equals(theme.getPrice());
     }
 }
